Hide experimental phantomjs wrapper
